updated-code

+++ server/app.py
@@ -1,75 +1,44 @@
-#!/usr/bin/env python
-
-from flask import Flask, jsonify, request, make_response
-from flask_migrate import Migrate
+from flask import Flask, jsonify, request
 from flask_restful import Api, Resource
-
+from flask_sqlalchemy import SQLAlchemy
 from models import db, Plant
 
 app = Flask(__name__)
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///plants.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///app.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.json.compact = False
-
-migrate = Migrate(app, db)
 db.init_app(app)
-
 api = Api(app)
 
-
-class Plants(Resource):
-
-    def get(self):
-        plants = [plant.to_dict() for plant in Plant.query.all()]
-        return make_response(jsonify(plants), 200)
-
-    def post(self):
-        data = request.get_json()
-
-        new_plant = Plant(
-            name=data['name'],
-            image=data['image'],
-            price=data['price'],
-            is_in_stock=data.get('is_in_stock', True)  # default to True if not provided
-        )
-
-        db.session.add(new_plant)
-        db.session.commit()
-
-        return make_response(new_plant.to_dict(), 201)
-
-
-api.add_resource(Plants, '/plants')
-
-
 class PlantByID(Resource):
-
     def get(self, id):
-        plant = Plant.query.filter_by(id=id).first_or_404()
-        return make_response(jsonify(plant.to_dict()), 200)
+        plant = db.session.get(Plant, id)
+        if not plant:
+            return {"error": "Plant not found"}, 404
+        return plant.to_dict(), 200
 
     def patch(self, id):
-        plant = Plant.query.filter_by(id=id).first_or_404()
-        data = request.get_json()
+        plant = db.session.get(Plant, id)
+        if not plant:
+            return {"error": "Plant not found"}, 404
 
+        data = request.get_json()
         if 'is_in_stock' in data:
             plant.is_in_stock = data['is_in_stock']
-        
-        db.session.commit()
-        
-        return make_response(jsonify(plant.to_dict()), 200)
+            db.session.commit()
+            return plant.to_dict(), 200
+
+        return {"error": "Invalid data"}, 400
 
     def delete(self, id):
-        plant = Plant.query.filter_by(id=id).first_or_404()
-        
+        plant = db.session.get(Plant, id)
+        if not plant:
+            return {"error": "Plant not found"}, 404
+
         db.session.delete(plant)
         db.session.commit()
-        
-        return make_response('', 204)
-
+        return '', 204
 
 api.add_resource(PlantByID, '/plants/<int:id>')
 
-
-if __name__ == '__main__':
-    app.run(port=5555, debug=True)
+with app.app_context():
+    db.create_all()
